requirements.txt
fastapi
uvicorn
pydantic
redis-om
python-jose[cryptography]
passlib
python-multipart